Removed all dependencies
All dependencies were removed because lua can convert ASCII codes to characters and vice versa by itself. Also rewrote some tables to not be ready tables. Fixed an issue with ask_input() returning a string instead of a number. Removed hint because this interpreter is now better and worth a try.